backend/requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.4,<4.5
psycopg2-binary>=2.9,<2.10
celery>=5.4,<5.5
django-celery-beat>=2.7,<2.8
redis>=5.0,<5.1
gunicorn>=22.0,<23.0
python-decouple>=3.8,<3.9
django-storages[s3]>=1.14,<1.15
boto3>=1.35,<1.36
Pillow>=10.4,<10.5
anthropic>=0.42,<1.0
resend>=2.5,<3.0
stripe>=11.4,<12.0
PyPDF2>=3.0,<4.0
python-docx>=1.0,<2.0
qrcode>=8.0,<9.0
